+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - The compiler can now discriminate between record access and module access
   for shadowed names
 - Check new project names against Erlang standard library modules to avoid clashes.
+- Formatter changed to treat lines with only spaces as empty.
 
 ## v0.12.0 - 2020-10-31